==> 01_sec1east_setup.R <==
## Source File for Broyles sec1east
##
## This script is for fitting the models used to predict yield and protein in sec1east.
## The models are saved as an .Rdata file. This script also generated performs the 
## analysis to generate an optimized N fertilizer map with the 2021 economic conditions. 
## The output table from this analysis is saved as an .Rdata file and the maps are saved 
## in the 'Outputs/OFPE' folder.

library(OFPE)

# import data and set up dat class

# fit/train the models

# export the models as .Rdata in general folder

# run the simulation with only 2021 economic data

# gather the simulation exported data to make maps

# calculate the total and mean for nitrogen applied, net-return, yield, and protein
# protein has 3 calcs: mean protein percent (%), mean protein as percent of yield (bu/ac), total protein as percent of yield (bu/ac)

# 1) map the optimal N fertilizer rates (total and mean N applied in caption)

# 2) map the predicted net-return (total and mean net-retun in caption)

# 3) map the predicted yield (total and mean yield in caption)

# 4) map the predicted protein (total and mean protein percent of yield and mean protein percent in caption)

# save all 4 maps to 'Strategies/OFPE'

# compile the mean and totals into table and save as .Rdata in 'Outputs/OFPE'
